Reorganized API functions
